update userID conversion in ValidateToken

+++ utils/jwt.go
@@ -48,7 +48,7 @@ func ValidateToken(token string) (int64, error) {
 	}
 
 	// email := claims["email"].(string)
-	userID := claims["userID"].(int64)
+	userID := int64(claims["userID"].(float64))
 
 	return userID, nil
 }
